requirements.txt
# Selenium testing dependencies
selenium            # WebDriver for browser automation
webdriver-manager   # Browser driver management
python-dotenv       # Environment variable management
